requirements.txt
pyaes==1.6.1
Pyrogram==2.0.57
PySocks==1.7.1
TgCrypto==1.2.3
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
frozenlist==1.3.1
idna==3.4
lxml==4.9.1
msspeech==3.7.5
multidict==6.0.2
python-docx==0.8.11
yarl==1.8.1
loguru==0.5.3
python-dotenv==0.20.0
